refactoring html thymeleaf to be usable in other sections
