+++ src/main/resources/templates/index.html
@@ -3,17 +3,11 @@
       xmlns="http://www.w3.org/1999/html" lang="en">
 <head th:replace="~{partials/head.html :: head('Home')}"><title>title</title></head>
 <body class="bg-color-neutral-dark">
-<!--<header th:replace="~{partials/header.html :: header}"></header>-->
-
-<main class="container ms-0">
+<main class="container-fluid ms-0">
     <div class="row gx-5">
-        <nav class="col-auto nav nav-pills d-flex flex-column bg-color-neutral-main px-4 py-5">
-            <a class="nav-link active" aria-current="page" href="#">Active</a>
-            <a class="nav-link" href="#">Link</a>
-            <a class="nav-link" href="#">Link</a>
-            <a class="nav-link disabled" aria-disabled="true">Disabled</a>
-        </nav>
-        <section class="col">
+        <nav th:replace="~{partials/primary-nav.html :: primary-nav}"></nav>
+
+        <section class="col-6 content">
             <h1 class="font-bold color-neutral-light my-3">Recent matches</h1>
             <ul class="nav nav-pills mb-3" id="pills-tab" role="tablist">
                 <li class="nav-item" role="presentation">
@@ -38,22 +32,22 @@
                         <div th:replace="~{partials/index/match.html :: match-all}"></div>
                     </th:block>
                 </div>
-                <div class="tab-pane fade" id="pills-aram" role="tabpanel" aria-labelledby="pills-aram-tab" tabindex="0">
+                <div class="tab-pane fade" id="pills-aram" role="tabpanel" aria-labelledby="pills-aram-tab" tabindex="1">
                     <th:block th:each="matchInfo : ${matchInfos}">
                         <div th:replace="~{partials/index/match.html :: match-aram}"></div>
                     </th:block>
                 </div>
-                <div class="tab-pane fade" id="pills-rift" role="tabpanel" aria-labelledby="pills-rift-tab" tabindex="0">
+                <div class="tab-pane fade" id="pills-rift" role="tabpanel" aria-labelledby="pills-rift-tab" tabindex="2">
                     <th:block th:each="matchInfo : ${matchInfos}">
                         <div th:replace="~{partials/index/match.html :: match-rift}"></div>
                     </th:block>
                 </div>
-                <div class="tab-pane fade" id="pills-arena" role="tabpanel" aria-labelledby="pills-arena-tab" tabindex="0">
+                <div class="tab-pane fade" id="pills-arena" role="tabpanel" aria-labelledby="pills-arena-tab" tabindex="3">
                     <th:block th:each="matchInfo : ${matchInfos}">
                         <div th:replace="~{partials/index/match.html :: match-arena}"></div>
                     </th:block>
                 </div>
-                <div class="tab-pane fade" id="pills-new" role="tabpanel" aria-labelledby="pills-new-tab" tabindex="0">
+                <div class="tab-pane fade" id="pills-new" role="tabpanel" aria-labelledby="pills-new-tab" tabindex="4">
                     <th:block th:each="matchInfo : ${matchInfos}">
                         <div th:replace="~{partials/index/match.html :: match-new}"></div>
                     </th:block>
@@ -63,33 +57,6 @@
 
         <aside th:replace="~{partials/index/friends-aside.html :: aside}"></aside>
     </div>
-
 </main>
-
-<footer th:replace="~{partials/footer.html :: footer}"></footer>
 </body>
 </html>
-
-
-<!--<div-->
-<!--        th:with="maxDamage=${T(java.lang.Math).max(matchInfo.getMatch().getTeam()[0].getTotalDamage(), matchInfo.getMatch().getTeam()[1].getTotalDamage())}"-->
-<!--&gt;-->
-<!--    <p th:text="${participant.getTotalDamageDealt()}"></p>-->
-<!--    <div class="progress-stacked">-->
-<!--        <div class="progress" role="progressbar" aria-label="Segment one" aria-valuenow="15" aria-valuemin="0" aria-valuemax="100"-->
-<!--             th:style="'width:'+${#numbers.formatDecimal((100 * participant.getPhysicalDamageDealt()/(maxDamage)), 0, 0) + '%'}"-->
-<!--        >-->
-<!--            <div class="progress-bar"></div>-->
-<!--        </div>-->
-<!--        <div class="progress" role="progressbar" aria-label="Segment two" aria-valuenow="30" aria-valuemin="0" aria-valuemax="100"-->
-<!--             th:style="'width:'+${#numbers.formatDecimal((100 * participant.getMagicDamageDealt()/(maxDamage)), 0, 0) + '%'}"-->
-<!--        >-->
-<!--            <div class="progress-bar bg-success"></div>-->
-<!--        </div>-->
-<!--        <div class="progress" role="progressbar" aria-label="Segment three" aria-valuenow="20" aria-valuemin="0"aria-valuemax="100"-->
-<!--             th:style="'width:'+${#numbers.formatDecimal((100 * participant.getTrueDamageDealt()/participant.getTotalDamageDealt()), 0, 0) + '%'}"-->
-<!--        >-->
-<!--            <div class="progress-bar bg-info"></div>-->
-<!--        </div>-->
-<!--    </div>-->
-<!--</div>-->
