+++ src/main/resources/templates/partials/index/item-icons.html
@@ -1,67 +1,25 @@
 <div th:fragment="items(title)"
      class="d-flex justify-content-between mt-1 gap-0 column-gap-1"
 >
-  <th:block th:if="${participant.getItem0() == 0}">
-    <div class="item-icon non-selectable opacity-75 bg-dark border border-1 border-secondary"></div>
-  </th:block>
-  <th:block th:unless="${participant.getItem0() == 0}">
-    <img th:src="@{${'/dragon_tail/13.14.1/img/item/' + participant.getItem0() + '.png'}}"
-         alt="player icon" class="item-icon non-selectable opacity-75 border border-1 border-secondary"
-    >
-  </th:block>
+  <div th:replace="~{partials/index/item-icons.html :: item(${participant.getItem0()})}"></div>
+  <div th:replace="~{partials/index/item-icons.html :: item(${participant.getItem1()})}"></div>
+  <div th:replace="~{partials/index/item-icons.html :: item(${participant.getItem2()})}"></div>
+  <div th:replace="~{partials/index/item-icons.html :: item(${participant.getItem3()})}"></div>
+  <div th:replace="~{partials/index/item-icons.html :: item(${participant.getItem4()})}"></div>
+  <div th:replace="~{partials/index/item-icons.html :: item(${participant.getItem5()})}"></div>
+  <div th:replace="~{partials/index/item-icons.html :: item(${participant.getItem6()})}"></div>
+</div>
 
-  <th:block th:if="${participant.getItem1() == 0}">
-    <div class="item-icon non-selectable opacity-75 bg-dark border border-1 border-secondary"></div>
-  </th:block>
-  <th:block th:unless="${participant.getItem1() == 0}">
-    <img th:src="@{${'/dragon_tail/13.14.1/img/item/' + participant.getItem1() + '.png'}}"
-         alt="player icon" class="item-icon non-selectable opacity-75 border border-1 border-secondary"
-    >
-  </th:block>
 
-  <th:block th:if="${participant.getItem2() == 0}">
+<div th:fragment="item(item)"
+     class="d-flex justify-content-between mt-1 gap-0 column-gap-1"
+>
+  <th:block th:if="${item == 0}">
     <div class="item-icon non-selectable opacity-75 bg-dark border border-1 border-secondary"></div>
   </th:block>
-  <th:block th:unless="${participant.getItem2() == 0}">
-    <img th:src="@{${'/dragon_tail/13.14.1/img/item/' + participant.getItem2() + '.png'}}"
-         alt="player icon" class="item-icon non-selectable opacity-75 border border-1 border-secondary"
-    >
-  </th:block>
-
-  <th:block th:if="${participant.getItem3() == 0}">
-    <div class="item-icon non-selectable opacity-75 bg-dark border border-1 border-secondary"></div>
-  </th:block>
-  <th:block th:unless="${participant.getItem3() == 0}">
-    <img th:src="@{${'/dragon_tail/13.14.1/img/item/' + participant.getItem3() + '.png'}}"
-         alt="player icon" class="item-icon non-selectable opacity-75 border border-1 border-secondary"
-    >
-  </th:block>
-
-  <th:block th:if="${participant.getItem4() == 0}">
-    <div class="item-icon non-selectable opacity-75 bg-dark border border-1 border-secondary"></div>
-  </th:block>
-  <th:block th:unless="${participant.getItem4() == 0}">
-    <img th:src="@{${'/dragon_tail/13.14.1/img/item/' + participant.getItem4() + '.png'}}"
-         alt="player icon" class="item-icon non-selectable opacity-75 border border-1 border-secondary"
-    >
-  </th:block>
-
-  <th:block th:if="${participant.getItem5() == 0}">
-    <div class="item-icon non-selectable opacity-75 bg-dark border border-1 border-secondary"></div>
-  </th:block>
-  <th:block th:unless="${participant.getItem5() == 0}">
-    <img th:src="@{${'/dragon_tail/13.14.1/img/item/' + participant.getItem5() + '.png'}}"
-         alt="player icon" class="item-icon non-selectable opacity-75 border border-1 border-secondary"
-    >
-  </th:block>
-
-  <th:block th:if="${participant.getItem6() == 0}">
-    <div class="item-icon non-selectable opacity-75 bg-dark border border-1 border-secondary"></div>
-  </th:block>
-  <th:block th:unless="${participant.getItem6() == 0}">
-    <img th:src="@{${'/dragon_tail/13.14.1/img/item/' + participant.getItem6() + '.png'}}"
+  <th:block th:unless="${item == 0}">
+    <img th:src="@{${'/dragon_tail/13.14.1/img/item/' + item + '.png'}}"
          alt="player icon" class="item-icon non-selectable opacity-75 border border-1 border-secondary"
     >
   </th:block>
 </div>
-
