+++ src/main/resources/templates/partials/index/position-icons.html
@@ -1,126 +1,48 @@
-<!--HORIZONTAL POSITION ICON SECTION-->
-<div th:fragment="horizontal(title)"
-        th:switch="${matchInfo.getMatch().getGameMode()}">
+<div th:fragment="position-icons(isVertical)"
+     th:switch="${matchInfo.getMatch().getGameMode()}"
+     class=""
+     th:classappend="${isVertical} ? 'bg-color-neutral-dark col-auto py-3 px-0' : ''"
+>
     <th:block th:case="CLASSIC">
-        <div th:replace="~{partials/index/position-icons.html :: classic-horizontal('')}"></div>
+        <div th:replace="~{partials/index/position-icons.html :: classic(${isVertical})}"></div>
     </th:block>
 
     <th:block th:case="ARAM">
-        <div th:replace="~{partials/index/position-icons.html :: aram-horizontal('')}"></div>
+        <div th:replace="~{partials/index/position-icons.html :: aram(${isVertical})}"></div>
     </th:block>
 </div>
 
-<div th:fragment="aram-horizontal(title)"
-     class="d-flex justify-content-between px-2 mb-2"
+<div th:fragment="aram(isVertical)"
+     class="d-flex justify-content-between align-items-center"
+     th:classappend="${isVertical} ? 'flex-column h-100' : 'px-2 mb-2'"
 >
-  <img th:src="@{${'/img/mid.png'}}"
-       alt="player position" class="position-icon non-selectable opacity-75"
-  >
-  <img th:src="@{${'/img/mid.png'}}"
-       alt="player position" class="position-icon non-selectable opacity-75"
-  >
-  <img th:src="@{${'/img/mid.png'}}"
-       alt="player position" class="position-icon non-selectable opacity-75"
-  >
-  <img th:src="@{${'/img/mid.png'}}"
-       alt="player position" class="position-icon non-selectable opacity-75"
-  >
-  <img th:src="@{${'/img/mid.png'}}"
-       alt="player position" class="position-icon non-selectable opacity-75"
-  >
+    <div th:replace="~{partials/index/position-icons.html :: position-icon('mid', ${isVertical})}"></div>
+    <div th:replace="~{partials/index/position-icons.html :: position-icon('mid', ${isVertical})}"></div>
+    <div th:replace="~{partials/index/position-icons.html :: position-icon('mid', ${isVertical})}"></div>
+    <div th:replace="~{partials/index/position-icons.html :: position-icon('mid', ${isVertical})}"></div>
+    <div th:replace="~{partials/index/position-icons.html :: position-icon('mid', ${isVertical})}"></div>
 </div>
 
-<div th:fragment="classic-horizontal(title)"
-     class="d-flex justify-content-between px-2 mb-2"
+<div th:fragment="classic(isVertical)"
+     class="d-flex justify-content-between align-items-center"
+     th:classappend="${isVertical} ? 'flex-column h-100' : 'px-2 mb-2'"
 >
-    <img th:src="@{${'/img/top.png'}}"
-         alt="player position top" class="position-icon non-selectable opacity-75"
-    >
-    <img th:src="@{${'/img/jgl.png'}}"
-         alt="player position jungle" class="position-icon non-selectable opacity-75"
-    >
-    <img th:src="@{${'/img/mid.png'}}"
-         alt="player position mid" class="position-icon non-selectable opacity-75"
-    >
-    <img th:src="@{${'/img/bot.png'}}"
-         alt="player position bot" class="position-icon non-selectable opacity-75"
-    >
-    <img th:src="@{${'/img/sup.png'}}"
-         alt="player position support" class="position-icon non-selectable opacity-75"
-    >
+    <div th:replace="~{partials/index/position-icons.html :: position-icon('top', ${isVertical})}"></div>
+    <div th:replace="~{partials/index/position-icons.html :: position-icon('jgl', ${isVertical})}"></div>
+    <div th:replace="~{partials/index/position-icons.html :: position-icon('mid', ${isVertical})}"></div>
+    <div th:replace="~{partials/index/position-icons.html :: position-icon('bot', ${isVertical})}"></div>
+    <div th:replace="~{partials/index/position-icons.html :: position-icon('sup', ${isVertical})}"></div>
 </div>
 
-<!--VERTICAL POSITION ICON SECTION-->
-<div  th:fragment="vertical(title)"
-      th:switch="${matchInfo.getMatch().getGameMode()}"
-      class="bg-color-neutral-dark col-auto py-3 px-0"
+<div th:fragment="position-icon(position, isVertical)"
+     class="d-flex justify-content-center align-items-center"
+     th:classappend="${isVertical} ? 'position-icon-lg-box mb-3' : ''"
 >
-    <th:block th:case="CLASSIC">
-        <div th:replace="~{partials/index/position-icons.html :: classic-vertical('')}"></div>
-    </th:block>
+    <img th:src="@{${'/img/' + position + '.png'}}"
+         th:alt="${'player position ' + position}"
+         class="non-selectable opacity-75"
+         th:classappend="${isVertical} ? 'position-icon-lg' : 'position-icon'"
+    >
 
-    <th:block th:case="ARAM">
-        <div th:replace="~{partials/index/position-icons.html :: aram-vertical('')}"></div>
-    </th:block>
 </div>
 
-<div th:fragment="classic-vertical(title)"
-     class="d-flex flex-column justify-content-between align-items-center h-100"
->
-    <div class="position-icon-lg-box d-flex justify-content-center align-items-center mb-3">
-        <img th:src="@{${'/img/top.png'}}"
-             alt="player position top" class="position-icon-lg non-selectable opacity-75"
-        >
-    </div>
-    <div class="position-icon-lg-box d-flex justify-content-center align-items-center mb-3">
-        <img th:src="@{${'/img/jgl.png'}}"
-             alt="player position jungle" class="position-icon-lg non-selectable opacity-75"
-        >
-    </div>
-    <div class="position-icon-lg-box d-flex justify-content-center align-items-center mb-3">
-        <img th:src="@{${'/img/mid.png'}}"
-             alt="player position mid" class="position-icon-lg non-selectable opacity-75"
-        >
-    </div>
-    <div class="position-icon-lg-box d-flex justify-content-center align-items-center mb-3">
-        <img th:src="@{${'/img/bot.png'}}"
-             alt="player position bot" class="position-icon-lg non-selectable opacity-75"
-        >
-    </div>
-    <div class="position-icon-lg-box d-flex justify-content-center align-items-center mb-3">
-        <img th:src="@{${'/img/sup.png'}}"
-             alt="player position support" class="position-icon-lg non-selectable opacity-75"
-        >
-    </div>
-</div>
-
-<div th:fragment="aram-vertical(title)"
-     class="d-flex flex-column justify-content-between align-items-center h-100"
->
-    <div class="position-icon-lg-box d-flex justify-content-center align-items-center mb-3">
-        <img th:src="@{${'/img/mid.png'}}"
-             alt="player position" class="position-icon-lg non-selectable opacity-75"
-        >
-    </div>
-
-    <div class="position-icon-lg-box d-flex justify-content-center align-items-center mb-3">
-        <img th:src="@{${'/img/mid.png'}}"
-             alt="player position" class="position-icon-lg non-selectable opacity-75"
-        >
-    </div>
-    <div class="position-icon-lg-box d-flex justify-content-center align-items-center mb-3">
-        <img th:src="@{${'/img/mid.png'}}"
-             alt="player position" class="position-icon-lg non-selectable opacity-75"
-        >
-    </div>
-    <div class="position-icon-lg-box d-flex justify-content-center align-items-center mb-3">
-        <img th:src="@{${'/img/mid.png'}}"
-             alt="player position" class="position-icon-lg non-selectable opacity-75"
-        >
-    </div>
-    <div class="position-icon-lg-box d-flex justify-content-center align-items-center mb-3">
-        <img th:src="@{${'/img/mid.png'}}"
-             alt="player position" class="position-icon-lg non-selectable opacity-75"
-        >
-    </div>
-</div>
