+++ src/main/resources/templates/partials/primary-nav.html
@@ -1,0 +1,7 @@
+  <nav th:fragment="primary-nav"
+       class="col-auto nav nav-pills d-flex flex-column bg-color-neutral-black px-5 py-5"
+  >
+    <a class="nav-link menu-btns p-0 mb-4 active" aria-current="page" id="matches-btn" href="/"></a>
+    <a class="nav-link menu-btns p-0 mb-4" id="challenges-btn" href="#"></a>
+    <a class=" menu-btns p-0 mb-4" id="friend-btn" href="#"></a>
+  </nav>
